replace b64
