+++ ui/app/components/b64-toggle.js
@@ -1,131 +1,80 @@
-import { equal } from '@ember/object/computed';
 import { isBlank } from '@ember/utils';
-import Component from '@ember/component';
-import { set, computed } from '@ember/object';
+import Component from '@glimmer/component';
+import { action } from '@ember/object';
+import { tracked } from '@glimmer/tracking';
 import { encodeString, decodeString } from 'vault/utils/b64';
 
-const B64 = 'base64';
-const UTF8 = 'utf-8';
-export default Component.extend({
-  tagName: 'button',
-  attributeBindings: ['type', 'data-test-transit-b64-toggle'],
-  type: 'button',
-  classNames: ['button', 'b64-toggle'],
-  classNameBindings: ['isInput:is-input:is-textarea'],
+/**
+ * @module B64Toggle2
+ * B64Toggle2 components are used to...
+ *
+ * @example
+ * ```js
+ * <B64Toggle2 @requiredParam={requiredParam} @optionalParam={optionalParam} @param1={{param1}}/>
+ * ```
+ * @param {object} requiredParam - requiredParam is...
+ * @param {string} [optionalParam] - optionalParam is...
+ * @param {string} [param1=defaultValue] - param1 is...
+ * @param {function} onChange - Function to handle the changing of the value passed in.
+ */
+export const B64 = 'base64';
+export const UTF8 = 'utf-8';
 
-  /*
-   * Whether or not the toggle is associated with an input.
-   * Also bound to `is-input` and `is-textarea` classes
-   * Defaults to true
-   *
-   * @public
-   * @type boolean
-   */
+export default class B64Toggle2 extends Component {
+  @tracked _value; // internal tracker of encoded value
+  @tracked lastEncoding = ''; // only becomes value once the action has been hit
 
-  isInput: true,
+  constructor() {
+    super(...arguments);
+  }
 
-  /*
-   * The value that will be mutated when the encoding is toggled
-   *
-   * @public
-   * @type string
-   */
-  value: null,
-
-  /*
-   * The encoding of `value` when the component is initialized.
-   * Defaults to 'utf-8'.
-   * Possible values: 'utf-8' and 'base64'
-   *
-   * @public
-   * @type string
-   */
-  initialEncoding: UTF8,
-
-  /*
-   * A cached version of value - used to determine if the input has changed since encoding.
-   *
-   * @private
-   * @type string
-   */
-  _value: '',
-
-  /*
-   * The current encoding of `value`.
-   * Possible values: 'utf-8' and 'base64'
-   *
-   * @private
-   * @type string
-   */
-  currentEncoding: '',
-
-  /*
-   * The encoding when we last mutated `value`.
-   * Possible values: 'utf-8' and 'base64'
-   *
-   * @private
-   * @type string
-   */
-  lastEncoding: '',
-
-  /*
-   * Is the value known to be base64-encoded.
-   *
-   * @private
-   * @type boolean
-   */
-  isBase64: equal('currentEncoding', B64),
-
-  /*
-   * Does the current value match the cached _value, i.e. has the input been mutated since we encoded.
-   *
-   * @private
-   * @type boolean
-   */
-  valuesMatch: computed('value', '_value', function () {
-    const { value, _value } = this;
-    const anyBlank = isBlank(value) || isBlank(_value);
-    return !anyBlank && value === _value;
-  }),
-
-  init() {
-    this._super(...arguments);
-    const initial = this.initialEncoding;
-    set(this, 'currentEncoding', initial);
-    if (initial === B64) {
-      set(this, '_value', this.value);
-      set(this, 'lastEncoding', B64);
-    }
-  },
-
-  didReceiveAttrs() {
-    this._super();
-    // if there's no value, reset encoding
-    if (this.value === '') {
-      set(this, 'currentEncoding', UTF8);
-      return;
-    }
-    // the value has changed after we transformed it so we reset currentEncoding
-    if (this.isBase64 && !this.valuesMatch) {
-      set(this, 'currentEncoding', UTF8);
-    }
-    // the value changed back to one we previously had transformed
+  get currentEncoding() {
+    // can only be two values: B64 or UTF8
+    // only set to B64 if lastEncoding was B64 and the valuesMatch
     if (this.lastEncoding === B64 && this.valuesMatch) {
-      set(this, 'currentEncoding', B64);
+      return B64;
     }
-  },
+    // otherwise in all cases it is UTF8 encoding
+    return UTF8;
+  }
 
-  click() {
-    let val = this.value;
+  get isBase64() {
+    return this.currentEncoding === B64 ? true : false;
+  }
+
+  get isInput() {
+    // isInput is either false or undefined so need to use equals instead of simple or statement
+    if (this.args.isInput === false) {
+      return false;
+    } else {
+      return true;
+    }
+  }
+
+  get valuesMatch() {
+    const anyBlank = isBlank(this.value);
+    return !anyBlank && this.args.value === this._value;
+  }
+
+  get value() {
+    return this._value || this.args.value;
+  }
+
+  @action
+  encodeDecodeValue() {
     const isUTF8 = this.currentEncoding === UTF8;
-    if (!val) {
+    if (!this.args.value) {
       return;
     }
-    let newVal = isUTF8 ? encodeString(val) : decodeString(val);
-    const encoding = isUTF8 ? B64 : UTF8;
-    set(this, 'value', newVal);
-    set(this, '_value', newVal);
-    set(this, 'lastEncoding', encoding);
-    set(this, 'currentEncoding', encoding);
-  },
-});
+    let newVal = isUTF8 ? encodeString(this.args.value) : decodeString(this.args.value);
+    let toggleEncodingType = isUTF8 ? B64 : UTF8;
+
+    this._value = newVal; // the encoded value
+    this.lastEncoding = toggleEncodingType;
+
+    if (this.args.onChange) {
+      // function passed to component to change the changed value
+      this.args.onChange(newVal);
+    }
+  }
+}
