+++ ui/app/templates/components/tool-hash.hbs
@@ -41,8 +41,7 @@
       </label>
       <div class="control">
         <Textarea id="input-hash" name="input" @value={{@input}} class="textarea" data-test-tools-input="hash-input" />
-        {{!-- <B64Toggle2 @isInput={{false}} @value={{@input}} @dataTestType="input" @onChange={{@onChange}} /> --}}
-        <B64Toggle @value={{@input}} @isInput={{false}} @dataTestType="input" />
+        <B64Toggle @isInput={{false}} @value={{@input}} @dataTestType="input" @onChange={{@onChange}} />
       </div>
     </div>
     <div class="field is-horizontal">
